refactor(math): mv log funcs in their own file
